Add keyboard shortcuts.
If you start typing a number,
and the input box doesn't already have the focus,
it will get the focus,
and the number you typed will replace what was already in there.

If you hit enter while in the input box,
That will set the value of the number,
Like hitting the `=` button.